refactor: Separate organisers social in components and add mask in mock
