+++ src/components/sections/organisers/organisers.scss
@@ -53,44 +53,11 @@
     }
 }
 
-.organisers-volkova {
-    &::after {
-        width: 272px;
-        height: 180px;
-        background-image: url('../../../core/mock/images/organisers/mask_volkova.png');
-    }
-}
-
-.organisers-pavlov {
-    &::after {
-        width: 132px;
-        height: 126px;
-        background-image: url('../../../core/mock/images/organisers/mask_pavlov.png');
-    }
-}
-
-.organisers-savelyev {
-    &::after {
-        width: 202px;
-        height: 126px;
-        background-image: url('../../../core/mock/images/organisers/mask_savelyev.png');
-    }
-}
-
-.organisers-palamarchuk {
-    &::after {
-        width: 206px;
-        height: 126px;
-        background-image: url('../../../core/mock/images/organisers/mask_palamarchuk.png');
-    }
-}
-
-.organisers-korotaev {
-    &::after {
-        width: 272px;
-        height: 160px;
-        background-image: url('../../../core/mock/images/organisers/mask_korotaev.png');
-    }
+.organisers-mask {
+    position: absolute;
+    bottom: -2;
+    left: 0;
+    pointer-events: none;
 }
 
 .organisers-card-img {
@@ -127,25 +94,4 @@
     min-width: 160px;
     max-width: 260px;
     padding: 12px 16px 8px;
-}
-
-.organisers-social {
-    position: absolute;
-    top: 0;
-    right: 0;
-    padding: 0 0 16px 16px;
-    background-color: var(--cl-background);
-}
-
-.organisers-social-item {
-    display: block;
-    color: var(--cl-text);
-    padding: 8px;
-    border: 1px solid var(--cl-text);
-    border-radius: 10px;
-
-    &:hover {
-        color: var(--cl-primary);
-        border-color: var(--cl-primary);
-    }
 }
