+++ src/core/mock/organisers.ts
@@ -1,8 +1,13 @@
 import juliaVolkova from "../mock/images/organisers/julia_volkova.png";
+import juliaVolkovaMask from "../mock/images/organisers/mask_volkova.png";
 import antonPavlov from "../mock/images/organisers/anton_pavlov.jpg";
+import antonPavlovMask from "../mock/images/organisers/mask_pavlov.png";
 import alexSavelyev from "../mock/images/organisers/alex_savelyev.png";
+import alexSavelyevMask from "../mock/images/organisers/mask_savelyev.png";
 import anastasiaPalamarchuk from "../mock/images/organisers/anastasia_palamarchuk.png";
+import anastasiaPalamarchukMask from "../mock/images/organisers/mask_palamarchuk.png";
 import aleksandrKorotaev from "../mock/images/organisers/aleksandr_korotaev.png";
+import aleksandrKorotaevMask from "../mock/images/organisers/mask_korotaev.png";
 
 export const organisers = [
     {
@@ -11,6 +16,7 @@ export const organisers = [
         id: "volkova",
         description: "Frontend Expert at Hyperskill and Mentor at HTML Academy",
         socialLink: "https://www.linkedin.com/in/julie-volkova/",
+        maskImage: juliaVolkovaMask,
     },
     {
         avatarSrc: antonPavlov.src,
@@ -18,6 +24,7 @@ export const organisers = [
         id: "pavlov",
         description: "DevOps",
         socialLink: "https://www.linkedin.com/in/antonepavlov/",
+        maskImage: antonPavlovMask,
     },
     {
         avatarSrc: alexSavelyev.src,
@@ -25,6 +32,7 @@ export const organisers = [
         id: "savelyev",
         description: "Frontend Developer",
         socialLink: "https://www.linkedin.com/in/vordgi/",
+        maskImage: alexSavelyevMask,
     },
     {
         avatarSrc: anastasiaPalamarchuk.src,
@@ -32,6 +40,7 @@ export const organisers = [
         id: "palamarchuk",
         description: "Tech Recruiter",
         socialLink: "https://www.linkedin.com/in/palamarchelo/",
+        maskImage: anastasiaPalamarchukMask,
     },
     {
         avatarSrc: aleksandrKorotaev.src,
@@ -39,5 +48,6 @@ export const organisers = [
         id: "korotaev",
         description: "Frontend Developer, SPB Frontend Organizer",
         socialLink: "https://www.linkedin.com/in/lekzd/",
+        maskImage: aleksandrKorotaevMask,
     },
 ];
